// src/common/enums/error-code.enum.ts
export enum ErrorCode {
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  USER_INACTIVE = 'USER_INACTIVE',
  USER_LOCKED = 'USER_LOCKED',
  DEPARTMENT_SELECTION_REQUIRED = 'DEPARTMENT_SELECTION_REQUIRED',
  INVALID_DEPARTMENT_ASSIGNMENT = 'INVALID_DEPARTMENT_ASSIGNMENT',
  ASSIGNMENT_INACTIVE = 'ASSIGNMENT_INACTIVE',
  ASSIGNMENT_EXPIRED = 'ASSIGNMENT_EXPIRED',
  DEPARTMENT_INACTIVE = 'DEPARTMENT_INACTIVE',
  ROLE_INACTIVE = 'ROLE_INACTIVE',
  PERMISSION_DENIED = 'PERMISSION_DENIED',
  ROLE_ACTION_NOT_ALLOWED = 'ROLE_ACTION_NOT_ALLOWED',
  SESSION_EXPIRED = 'SESSION_EXPIRED',
  SESSION_REVOKED = 'SESSION_REVOKED',
  PERMISSION_VERSION_CHANGED = 'PERMISSION_VERSION_CHANGED',
  LAST_SUPER_ADMIN_PROTECTED = 'LAST_SUPER_ADMIN_PROTECTED',
  DUPLICATE_USERNAME = 'DUPLICATE_USERNAME',
  DUPLICATE_EMAIL = 'DUPLICATE_EMAIL',
  DUPLICATE_DEPARTMENT_CODE = 'DUPLICATE_DEPARTMENT_CODE',
  DUPLICATE_MENU_CODE = 'DUPLICATE_MENU_CODE',
  DUPLICATE_MENU_PATH = 'DUPLICATE_MENU_PATH',
}